% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDS.R
\name{rDS}
\alias{rDS}
\title{A read Data Series}
\usage{
rDS(fileNameList)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
rDS()
}
\keyword{cats}

